dom/filehandle/test/test_overlapping_filehandles.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    for each (let fileStorage in fileStorages) {
      let request = getMutableFile(fileStorage.key, "test.txt");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      let mutableFile = event.target.result;
      mutableFile.onerror = errorHandler;

      for (let i = 0; i < 50; i++) {
        let stepNumber = 0;

        request = mutableFile.open("readwrite").append("string1");
        request.onsuccess = function(event) {
          is(stepNumber, 1, "This callback came first");
          stepNumber++;
          event.target.fileHandle.oncomplete = grabEventAndContinueHandler;
        }

        request = mutableFile.open("readwrite").append("string2");
        request.onsuccess = function(event) {
          is(stepNumber, 2, "This callback came second");
          stepNumber++;
          event.target.fileHandle.oncomplete = grabEventAndContinueHandler;
        }

        request = mutableFile.open("readwrite").append("string3");
        request.onsuccess = function(event) {
          is(stepNumber, 3, "This callback came third");
          stepNumber++;
          event.target.fileHandle.oncomplete = grabEventAndContinueHandler;
        }

        stepNumber++;
        yield undefined; yield undefined; yield undefined;;

        is(stepNumber, 4, "All callbacks received");
      }
    }

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
